Add user sign up and login pages
Since having the ability to sign in is critical for our project. I added in user creation using passport and mongob. We didn't have any models for a user in the database, so I created that too(can be found in model/models.js). At the moment, this doesn't allow for any kind of session, it just tells the user that their login did succeed. Adding in a jsonwebtoken to allow for continual access is a top priority.