Added in wait time reporting
Wait time reporting is potentially the most critical feature, so this adds this in. In particular, this adds in the endpoints /user/recordTime, which takes a resturaunt(see the array at the top of index.js for values) and a time in seconds. Like any /user endpoint, the user needs to be authenticated by JWT. We also provide two endpoints to actually read these values, which are /predictions(reads in the historic times for the day and some prediction) as well as /waitTime, which gets the current average wait time.