Added in prediction/longterm data storage of values
Prediction is one of the killer features in our system, so this seeks to add it in. This adds no new endpoints from the previous commit, but adds functionality to several that did nothing before. As a sidenote, testing code that needs to fire at a specific time is difficult, so I can't guarantee that the code in there is perfect. To show that I did still do some testing, as well as give people the opportunity to play around with the system with some commented out test endpoints. I think these will probably get deleted later, but I'm writing this at 3:37AM on thanksgiving, so please cut me some slack for the braindead commit.