Fixed login issues by resolving preflight reuqest
Our previous way of handling cors was simple, just add the headers on to the request. This worked for simple requests, but some of our advanced stuff required a preflight, which we didn't handle. This handles the preflight response. This also fixes some issues with fetch not allowing a body by changing those get requests to take url params.